Allow customizable contribution instructions link
If the link is not set in a project's config, it defaults to linking them the
Rust contrib guide.

If the config specifies a link, it will be used in the new contributor welcome
message instead fo the Rust guide.

This is necessary because most of the smaller repos have a workflow that runs
tests with Travis and doesn't involve Bors, so pointing their contributors at
the main Rust workflow documentation is misleading.