Fix unavailable diff when working with private repos
Before this commit, highfive fetched the PR diff from the URL provided
by the webhook, which is `https://github.com/org/repo/pull/42.diff`.

That works fine for public repos, but for private repos that URL is
private, and highfive can't fetch it. This commit changes the code to
properly fetch the diff from the API.

https://developer.github.com/v3/media/#commits-commit-comparison-and-pull-requests