Using commit search to identify new contributors
The current approach to checking for new contributors does not work
in repositories with more than 500 contributors because the
contributors list API endpoint only associates the first 500
committer emails with GitHub usernames. See #41 for more
explanation.

This commit modifies the new contributor check to use the commit
search API endpoint. A drawback of this approach is that (at least
currently) that endpoint does not provide useful information for
fork repositories, which is why the new contributor check has a
condition on whether the repository is a fork.