Remove ollie from the rustdoc review queue
They haven't been actively reviewing in 6+ months; in the past I was reassigning their PRs to mine, but I don't want to keep doing that. They're still a member of the team and can raise blocking concerns on RFCs/FCPs if they wish.

cc @rust-lang/rustdoc - this leaves only two people on the review queue, so it would be great if a few of you could join the rotation.