Adjust reviewer selection to work on unprefixed group selection
Previously, `r? compiler-team` or `r? pnkfelix` would not be matched by the
regexes, which meant that they'd be silently ignored. `r?
rust-lang/compiler-team` already worked, though. Note that if a user and team
name overlap, then `r? @foo` will always map to foo the user, while `r? foo`
will map to the team.

Since we generally don't want the `@` prefix on team-wide pings, since it
subscribes a bunch of people for no reason, it's nice to avoid necessarily
typing the prefix as well. Users in practice do invoke the command in that way.