Mention @obi1kenobi for rustdoc JSON types changes
I'm one of the maintainers of `cargo-semver-checks`, and as mentioned [in Zulip](https://rust-lang.zulipchat.com/#narrow/stream/266220-rustdoc/topic/Long.20Term.20Rustdoc.20JSON.20Stability) it's very useful to me to know about upcoming rustdoc JSON types changes so I can be ready to prepare new `cargo-semver-checks` releases compatible with the new JSON format.

I hope this is a reasonable and appropriate mechanism for that, and not an imposition on the team. Apologies for any mistakes in the file — this is the first time I've seen the internals of highfive.